requirements.txt
aiohttp
discord.py
firebase_admin
pokepy
